+++ tests/rates_providers/coinlayer-provider.test.js
@@ -1,35 +1,10 @@
 const CoinlayerProvider = require('../../rates_providers/coinlayer-provider');
 const coinlayer = require('../../http/axios-coinlayer');
+const { successResponse, ratesInfo, failResponse } = require('./test-data');
 
 jest.mock('../../http/axios-coinlayer');
 
 describe('CoinlayerProvider: fetchBtcToUahRate method', () => {
-  const successResponse = {
-    data: {
-      success: true,
-      terms: 'https://coinlayer.com/terms',
-      privacy: 'https://coinlayer.com/privacy',
-      timestamp: 1629755646,
-      target: 'UAH',
-      rates: { BTC: 1321519.72395 },
-    },
-  };
-
-  const ratesInfo = {
-    timestamp: 1629755646,
-    target: 'UAH',
-    rates: { BTC: 1321519.72395 },
-  };
-
-  const failResponse = {
-    success: false,
-    error: {
-      code: 101,
-      type: 'invalid_access_key',
-      info: 'You have not supplied a valid API Access Key. [Technical Support: support@apilayer.com]',
-    },
-  };
-
   test('should return basic BTC to UAH rates info', async () => {
     coinlayer.get.mockReturnValue(successResponse);
